revert some changes to fix bug 54
commit f6e807687e causes for some reason
a strange linking error/build break.

I tracked down the error to binfo/envsetup.sh
and my quess is that it's somehow caused by the
variable value changed to be in quotas and then
passed as a parameter to hipcc for example.

Hard to debug so reverting now some part of the code while
trying to keep many changes and comments in.

With this version the suitesparse builds and links again.

fixes: https://github.com/lamikr/rocm_sdk_builder/issues/54

Signed-off-by: Mika Laitio <lamikr@gmail.com>