add support for building single binfo files
- add new ./babs.sh -b binfo/myapp.binfo option

This new option will allow specifying binfo file as a parameter
for -b command. If sources are not yet downloaded, it will also
download them before building the application and then apply
patches if they exist in patches directory.

Note that this command does not check any external dependencies
to other applications or libraries.

fixes: https://github.com/lamikr/rocm_sdk_builder/issues/23

Signed-off-by: Mika Laitio <lamikr@gmail.com>