Build triton with moderate number of cores
Like aotriton, triton thinks building with twice the number of workers as the system has cores is a good idea. It's not, and doing so makes the build *slower* than it needs to be.

Not sure if we should use `MODERATE` or `SAFE` here; `MODERATE` works on my 12-core, 96 GB system but still consumes a lot of memory that lesser systems might not have.