better amdsmi fix for double free corruption (!prev) on pytorch 2.4.0 exit
pytorch 2.4.0 showed always on exit the error message
about double free or corruption. This happened even with
very simple cuda applications on rocm environment.

Error happened because const chars on map would be
tried to be released. Fix is to change the const chars to std::string
in map structure that holds message strings.

Signed-off-by: Mika Laitio <lamikr@gmail.com>