fix DeepSpeed ubuntu 22.04 build
error: can't copy 'deepspeed/accelerator': doesn't exist or not a regular file

This error happens on setup.py in intall phase and reason
is that the deepspeed/accelerator, deepspeed/ops/csrc and deepspeed/ops/op_builder
are symlinks in deepseed and that causes problem on setup.py on
Ubuntu 22.04 during the install phase. Error message is that
deepspeed/accelerator does not exist or is not a file.

Fix is to delete symlinks and use cp -ax instead to copy the directories
This is better than adding a patch as it quarantees that files are always up to date
when deepspeed version changes.

fixes: https://github.com/lamikr/rocm_sdk_builder/issues/163

Signed-off-by: Mika Laitio <lamikr@gmail.com>