add support for debian based LMDE6
GCC 12.20 on LDME6 has some bug or is even more strict
for handling the warnings as an error than the gcc13/14
on fedora. I had earlier the

CMAKE_CXX_FLAGS -Wno-error=maybe-uninitialized
but for LDME6, i need to add another flag also:
$CMAKE_CXX_FLAGS -Wno-error=maybe-uninitialized -Wno-error=uninitialized

These warnings are fixed in newer fbgemm which is one
sub-project that pytorch uses, but for current pytorch stable
version this is now better to fix just by not treating gcc
warnings as an error.

fixes: https://github.com/lamikr/rocm_sdk_builder/issues/175

Signed-off-by: Mika Laitio <lamikr@gmail.com>